requirements.txt
os
shutil
random
torch
torchvision
numpy
pillow
matplotlib
time
argparse
torchsummary
scikit-learn